refactored messaging
