+++ main/src/study_tinder/messaging/MessageHistory.java
@@ -1,4 +1,6 @@
-package study_tinder;
+package study_tinder.messaging;
+
+import study_tinder.User;
 
 import java.util.ArrayList;
 import java.util.List;
